Building payoff and withdral mechanism.
The Infinity Pool aggregates capital and deploys it towards productive endeavors that generate exponential economic growth. That revenue thus generated is then distributed to DAO members that contributed to creating it. This commit adds a lot of the functionality enabling the  withdrawal of the share of pool that DAO members have earned by funding and governing the DAO's operations.